+++ controllers/auth/testController.js
@@ -1,0 +1,10 @@
+// for testing if only admin can access the route
+
+export const testController = (req, res) => {
+  try {
+    res.send("Protected Routes");
+  } catch (error) {
+    console.log(error);
+    res.send({ error });
+  }
+};
