+++ jest.backend.config.js
@@ -6,7 +6,7 @@ export default {
   testEnvironment: "node",
 
   // which test to run
-  testMatch: ["<rootDir>/controllers/*.test.js"],
+  testMatch: ["<rootDir>/controllers/auth/*.test.js"],
 
   // jest code coverage
   collectCoverage: true,
